fix database application.properties

+++ ITing-backend/src/main/java/com/ITing/JobPortal/job/dto/JobSearchRequest.java
@@ -4,14 +4,16 @@ import com.iting.jobportal.job.entity.enums.ExperienceLevel;
 import com.iting.jobportal.job.entity.enums.JobType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class JobSearchRequest {
     private String keyword;         // Tìm theo từ khóa (position, description)
     private String location;        // Lọc theo địa điểm
     private JobType jobType;        // Lọc theo loại công việc
     private ExperienceLevel experienceLevel; // Lọc theo cấp bậc
-    private Long minSalary;         // Lọc theo mức lương tối thiểu
-    private Long maxSalary;         // Lọc theo mức lương tối đa
+    private BigDecimal minSalary;         // Lọc theo mức lương tối thiểu
+    private BigDecimal maxSalary;         // Lọc theo mức lương tối đa
     private Long companyId;         // Lọc theo công ty
     private String techRequired;    // Lọc theo công nghệ
